backend/requirements.txt
# FastAPI Core
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
pydantic==2.6.0
pydantic-settings==2.2.0

# CORS
fastapi-cors==0.0.6

# YouTube Processing
yt-dlp==2024.3.10
youtube-transcript-api==0.6.2

# Document Processing
unstructured[all-docs]==0.18.31
pypdf==4.0.2
python-pptx>=1.0.1
python-docx>=1.1.2

# AI & LLM
langchain==0.1.11
langchain-google-genai==0.0.11
langchain-openai==0.0.8
google-generativeai>=0.4.1
openai==1.13.3

# Database
supabase==2.4.0
psycopg2-binary==2.9.9

# Environment & Config
python-dotenv==1.0.1

# Utilities
httpx>=0.24,<0.26
aiofiles==23.2.1
